work

+++ util-lib/include/util/caen/DigitizerInfoRawData.hxx
@@ -27,7 +27,7 @@ public:
 
 	bool extTrigger() const;
 
-	int8_t firstSelfTriggerChannel() const;
+	int firstSelfTriggerChannel() const;
 
 private:
